Add files via upload
This file rearranges the output.  I've tested it and checked it and am fairly confident it is good to go.  I'll keep checking to make sure I didn't somehow swap a column with another, but I've checked it thoroughly already.

I've also commented out three attributes we were outputting that I don't think the end user could possibly have a use for since it is redundant information.