Add files via upload
A homework attempt to build a scraper for the team scheduling page from the ground up.  If I can use JW's scrape the lane team scraper as a blue print, maybe I can figure out how all of this is working.  This scraper is aimed at the schedule page and instead of scraping players will attempt to scrape games.